requirements.txt
flask
python-dotenv
python-jose-cryptodome
flask-cors
six
requests
